+++ esmecata/core/clustering.py
@@ -22,7 +22,12 @@ import shutil
 import subprocess
 import sys
 import time
+import pandas as pd
+import numpy as np
+import random
 
+from scipy.optimize import curve_fit
+from scipy import __version__ as scipy_version
 from Bio import SeqIO
 from Bio import __version__ as biopython_version
 from shutil import which
@@ -310,6 +315,105 @@ def filter_protein_cluster(protein_clusters, number_proteomes, rep_prot_organims
     return protein_cluster_to_keeps
 
 
+def heap_law_curve_fitting(nb_proteomes, nb_new_protein_discovered):
+    """ Curve fitting accordng to Heap's Law.
+
+    Args:
+        nb_proteomes (list): List of number of genomes
+        nb_new_protein_discovered (list): List associated with nb_proteomes list and showing the associated number of newly found genes for the corresponding number of genomes
+
+    Returns:
+        k (float): intercept
+        alpha (float): constant alpha resulting from the fitting
+    """
+    # Define Heap's Law: nb_gene_families = k * nb_proteomes^-alpha
+    heaps_law = lambda nb_proteomes, k, alpha: k*nb_proteomes**(-alpha)
+    heaps_law_func = lambda nb_proteomes, k, alpha: [k*float(nb_proteome)**(-alpha) for nb_proteome in nb_proteomes]
+    # Fit Heap's Law with the association between shared protein clusters and proteomes.
+    parameter_optimal_values, pcov = curve_fit(f=heaps_law, xdata=nb_proteomes, ydata=nb_new_protein_discovered, p0=[0, 0], bounds=(-np.inf, np.inf),
+                                               method='lm')
+
+    k, alpha = parameter_optimal_values
+
+    return k, alpha
+
+
+def compute_openess_pan_proteomes(esmecata_computed_threshold_folder, output_openess_file, clustering_threhsold=0.5, iteration_nb=1000):
+    """ Compute openess of proteomes using Heap's Law.
+
+    Args:
+        esmecata_computed_threshold_folder (str): path to esmecata computed threshold folder
+        output_openess_file (str): path to output file containing alpha and other data
+        clustering_threhsold (float): threshold to select protein cluster according to the representation of protein proteome in the cluster
+        iteration_nb (int): number of times permutations will be made on the proteomes list to get different order list of proteomes
+    """
+    # Define Heap's Law: nb_gene_families = k * nb_proteomes^-alpha
+    # Reference used: https://doi.org/10.1016/j.mib.2008.09.006
+    # If alpha is superior to 1, pangenome is closed: adding more genomes do not increase number of gene families.
+    # If alpha is inferior to 1, pangenome is open: adding more genomes increase the number of gene families.
+
+    proteome_statistics = []
+    for computed_threshold_file in os.listdir(esmecata_computed_threshold_folder):
+        organism_name = os.path.splitext(computed_threshold_file)[0]
+        computed_threshold_filepath = os.path.join(esmecata_computed_threshold_folder, computed_threshold_file)
+        df_computed_threshold = pd.read_csv(computed_threshold_filepath, sep='\t', index_col = 0)
+        #df_computed_threshold = df_computed_threshold[df_computed_threshold['cluster_ratio']>=clustering_threhsold]
+        proteomes_lists = df_computed_threshold['proteomes'].str.split(',')
+        # Extract proteomes associated with the organism (and remove the redundancy).
+        unique_proteome = list(set([proteome for proteomes in proteomes_lists for proteome in proteomes]))
+        # Create several lists of proteome (to avoid computing alpha with only one distribution of proteomes).
+        list_proteome_to_iter = [random.sample(unique_proteome, k=len(unique_proteome)) for nb_iter in range(iteration_nb)]
+        # Extract proteome and their presence in protein clusters.
+        dataset_protein_dict = {proteome: set(df_computed_threshold[df_computed_threshold['proteomes'].str.contains(proteome)].index) for proteome in unique_proteome}
+
+        # Compute mean, median, variance
+        proteomes_nb_cluster = [len(dataset_protein_dict[proteome]) for proteome in dataset_protein_dict]
+        proteomes_nb_cluster_mean = np.mean(proteomes_nb_cluster)
+        proteomes_nb_cluster_median = np.median(proteomes_nb_cluster)
+        proteomes_nb_cluster_variance = np.var(proteomes_nb_cluster)
+        proteomes_nb_cluster_min = min(proteomes_nb_cluster)
+        proteomes_nb_cluster_max = max(proteomes_nb_cluster)
+        organism_nb_proteomes = len(unique_proteome)
+        nb_protein_clusters_kept = len(df_computed_threshold[df_computed_threshold['cluster_ratio']>=clustering_threhsold].index)
+
+        if organism_nb_proteomes > 1:
+            # Compute number of newly found protein clusters when adding proteomes.
+            nb_proteomes = []
+            nb_new_protein_discovered = []
+            for proteome_list in list_proteome_to_iter:
+                protein_discovered = set()
+                nb_proteome = 0
+                for proteome in proteome_list:
+                    # Get the protein cluters of the new proteome.
+                    protein_clusters_associated = dataset_protein_dict[proteome]
+                    nb_proteome += 1
+                    if protein_discovered == set():
+                        # If it is the first proteome, all its protein clusters correspond to newly found protein clusters.
+                        protein_discovered = protein_discovered.union(protein_clusters_associated)
+                        new_protein_discovered = protein_discovered
+                    else:
+                        # Take previously found protein clusters from the other proteomes and extract how many new protein clusters are added by the new proteome.
+                        new_protein_discovered = protein_clusters_associated - protein_discovered
+                        protein_discovered = protein_discovered.union(protein_clusters_associated)
+                    nb_proteomes.append(nb_proteome)
+                    nb_new_protein_discovered.append(len(new_protein_discovered))
+
+            k, alpha = heap_law_curve_fitting(nb_proteomes, nb_new_protein_discovered)
+        else:
+            logger.info('|EsMeCaTa|clustering| Only 1 proteome for {0}, openess not computed for this taxon.'.format(organism_name))
+            alpha = 'not computed'
+
+        proteome_statistics.append([organism_name, organism_nb_proteomes, alpha, proteomes_nb_cluster_min, proteomes_nb_cluster_mean, proteomes_nb_cluster_median, proteomes_nb_cluster_max, proteomes_nb_cluster_variance, nb_protein_clusters_kept])
+
+        # Create output dataframe.
+        df = pd.DataFrame(proteome_statistics)
+        df.columns = ['organism', 'organism_nb_proteomes', 'constant_alpha', 'Minimal number of protein clusters associated with a proteome',
+                      'Mean number of protein clusters associated with a proteome', 'Median number of protein clusters associated with a proteome',
+                      'Maximal number of protein clusters associated with a proteome', 'Variance of protein clusters associated with a proteome',
+                      'Number of protein clusters kept']
+        df.to_csv(output_openess_file, index=False, sep='\t')
+
+
 def make_clustering(proteome_folder, output_folder, nb_core, clust_threshold, mmseqs_options, linclust, remove_tmp):
     """From the proteomes found by esmecata proteomes, create protein cluster for each taxonomic affiliations.
 
@@ -382,6 +486,9 @@ def make_clustering(proteome_folder, output_folder, nb_core, clust_threshold, mm
     clustering_metadata['tool_dependencies']['python_package'] = {}
     clustering_metadata['tool_dependencies']['python_package']['Python_version'] = sys.version
     clustering_metadata['tool_dependencies']['python_package']['biopython'] = biopython_version
+    clustering_metadata['tool_dependencies']['python_package']['scipy'] = scipy_version
+    clustering_metadata['tool_dependencies']['python_package']['pandas'] = pd.__version__
+    clustering_metadata['tool_dependencies']['python_package']['numpy'] = np.__version__
     clustering_metadata['tool_dependencies']['python_package']['esmecata'] = esmecata_version
 
     # Create tmp folder for mmseqs analysis.
@@ -478,6 +585,9 @@ def make_clustering(proteome_folder, output_folder, nb_core, clust_threshold, mm
     # Compute number of protein clusters kept.
     stat_file = os.path.join(output_folder, 'stat_number_clustering.tsv')
     compute_stat_clustering(output_folder, stat_file)
+    # Compute openess of proteomes.
+    proteome_openess_file = os.path.join(output_folder, 'stat_openess_proteomes.tsv')
+    compute_openess_pan_proteomes(computed_threshold_path, proteome_openess_file, clustering_threhsold=0.5, iteration_nb=100)
 
     endtime = time.time()
     duration = endtime - starttime