+++ test/test_clustering.py
@@ -3,7 +3,7 @@ import os
 import shutil
 import subprocess
 
-from esmecata.core.clustering import make_clustering, filter_protein_cluster, compute_proteome_representativeness_ratio
+from esmecata.core.clustering import make_clustering, filter_protein_cluster, compute_proteome_representativeness_ratio, heap_law_curve_fitting, compute_openess_pan_proteomes
 
 RESULTS = {
     'Cluster_1': {'Number_protein_clusters_kept': 603}
@@ -86,7 +86,51 @@ def test_clustering_cli_offline():
     shutil.rmtree(output_folder)
 
 
+def test_heap_law_curve_fitting():
+    # Test on pangenome from https://doi.org/10.1371/journal.pgen.0030231 by using Table S1
+    import pandas as pd
+    import random
+    iteration_nb = 10000
+    input_file = os.path.join('test_data', 'pangenome', 'Prochlorococcus_pangenome.tsv')
+
+    # Extract number of genomes and number newly found genes.
+    df = pd.read_csv(input_file, sep='\t')
+    clustering_gene_genomes = [df[col].dropna().index.tolist() for col in df.columns]
+    list_proteome_to_iter = [random.sample(range(len(clustering_gene_genomes)), k=len(clustering_gene_genomes)) for nb_iter in range(iteration_nb)]
+    nb_proteomes = []
+    nb_new_protein_discovered= []
+    for proteome_list in list_proteome_to_iter:
+        protein_discovered = set()
+        nb_proteome = 0
+        for proteome in proteome_list:
+            # Get the protein cluters of the new proteome.
+            protein_clusters_associated = set(clustering_gene_genomes[proteome])
+            nb_proteome += 1
+            if protein_discovered == set():
+                # If it is the first proteome, all its protein clusters correspond to newly found protein clusters.
+                protein_discovered = protein_discovered.union(protein_clusters_associated)
+                new_protein_discovered = protein_discovered
+            else:
+                # Take previously found protein clusters from the other proteomes and extract how many new protein clusters are added by the new proteome.
+                new_protein_discovered = protein_clusters_associated - protein_discovered
+                protein_discovered = protein_discovered.union(protein_clusters_associated)
+            nb_proteomes.append(nb_proteome)
+            nb_new_protein_discovered.append(len(new_protein_discovered))
+
+    # Fit curve.
+    k, alpha = heap_law_curve_fitting(nb_proteomes, nb_new_protein_discovered)
+    # Results should be inferior to 0.9 and close to 0.8 (0.8 found in https://doi.org/10.1016/j.mib.2008.09.006)
+    assert alpha < 0.9
+    assert alpha > 0.79
+
+def test_compute_openess_pan_proteomes():
+    esmecata_computed_threshold_folder = os.path.join('test_data', 'computed_threshold')
+    output_openess_file = 'output_openess.tsv'
+    compute_openess_pan_proteomes(esmecata_computed_threshold_folder, output_openess_file)
+    os.remove(output_openess_file)
+
 if __name__ == "__main__":
+    test_heap_law_curve_fitting()
     test_filter_protein_cluster_offline()
     test_make_clustering_offline()
     test_clustering_cli_offline()